download working

+++ Source/Client/client_main.cpp
@@ -84,10 +84,10 @@ int main(int argc, char** argv) {
                 continue;
             }
 
-            /*if(file_already_existing(filename)){
+            if(file_already_existing(filename)){
                 cout<<"Warning, file already exists and it will be overwritten"<<endl;
                 delete_file(filename);
-            }*/
+            }
 
             if(!handle_download(client_socket, identity, filename)){
                 cerr << "Error in contacting the server, disconnecting!" << endl;
